+++ klibs/tests/test_KLStructure.py
@@ -4,8 +4,11 @@ import pytest
 import random
 from collections import Counter
 
-from klibs.KLStructure import FactorSet
-from klibs.KLTrialFactory import _generate_blocks
+import klibs.KLParams as P
+from klibs.KLStructure import FactorSet, Block
+from klibs.KLTrialFactory import _generate_blocks, _load_structure, _parse_structure
+
+from conftest import create_tempfile
 
 
 class TestFactorSet(object):
@@ -135,3 +138,145 @@ def test_generate_blocks():
     assert block[0]['soa'] == 200 and block[0]['cue_loc'] == 'none'
     assert block[1]['soa'] == 0 and block[1]['easy_trial'] == True
     assert block[2]['soa'] == 800 and block[2]['cue_loc'] == 'right'
+
+
+
+class TestBlock(object):
+
+    def test_init(self):
+
+        factors = FactorSet({
+            'cue_loc': ['left', 'right', 'none'],
+            'easy_trial': [True, False],
+        })
+
+        # Test simple initialization
+        P.trials_per_block = 0
+        tst = Block(factors)
+        assert tst.trialcount == 6
+        assert tst.practice == False
+        assert tst.label == None
+
+        # Test initialization with dict
+        tst = Block({'soa': [200, 800], 'target_loc': ['L', 'R']})
+        assert tst.trialcount == 4
+        assert isinstance(tst._factors, FactorSet)
+
+        # Test defaulting trial count to P.trials_per_block
+        P.trials_per_block = 30
+        tst = Block(factors)
+        assert tst.trialcount == 30
+
+        # Test custom trial counts:
+        tst = Block(factors, trials=36)
+        assert tst.trialcount == 36
+
+        # Test labels and practice flags
+        tst = Block(factors, label='endo', practice=True)
+        assert tst.label == 'endo'
+        assert tst.practice == True
+
+    
+    def test_get_trials(self):
+
+        factors = FactorSet({
+            'cue_loc': ['left', 'right', 'none'],
+            'easy_trial': [True, False],
+        })
+
+        # Test generating trials with specified trial count
+        tst = Block(factors, trials=30)
+        trials = tst.get_trials()
+        assert len(trials) == 30
+        assert isinstance(trials[0], dict)
+        assert 'cue_loc' in list(trials[0].keys())
+
+        # Test partial shuffling
+        random.seed(308053045)
+        trials = tst.get_trials()
+        easy_count = 0
+        left_count = 0
+        for trial in trials[:6]:
+            easy_count += int(trial['easy_trial'] == True)
+            left_count += int(trial['cue_loc'] == 'left')
+        assert easy_count == 3
+        assert left_count == 2
+
+
+    def test_factors(self):
+
+        tst = Block({'a': [1, 2], 'b': [3, 4], 'c': [5, 6]})
+        assert tst.factors == ['a', 'b', 'c']
+
+
+def test_load_structure():
+    # NOTE: Move to KLTrialFactory tests once created
+    header = "from klibs.KLStructure import FactorSet, Block"
+
+    # Test loading structure
+    tmp = create_tempfile([
+        header, "",
+        "structure = [",
+        "    Block({}, label='a', trials=10),",
+        "    Block({}, label='B', trials=20)",
+        "]"
+    ])
+    tst = _load_structure(tmp)
+    assert len(tst) == 2
+    assert isinstance(tst[0], Block)
+
+    # Test loading missing structure
+    tmp = create_tempfile([
+        header, "",
+        "exp_factors = FactorSet({})"
+    ])
+    tst = _load_structure(tmp)
+    assert not tst
+
+    # Test loading empty structure
+    tmp = create_tempfile([
+        header, "",
+        "exp_factors = FactorSet({})",
+        "",
+        "structure = []",
+    ])
+    tst = _load_structure(tmp)
+    assert not tst
+
+
+def test_parse_structure():
+    # NOTE: Move to KLTrialFactory tests once created
+    exp_factors = {'fac1': ['a', 'b', 'c'], 'fac2': [True, False]}
+    tst = [
+        Block(exp_factors, label='a', trials=10, practice=True),
+        Block(exp_factors, label='b', trials=20)
+    ]
+    blocks = _parse_structure(tst, exp_factors)
+    assert len(blocks) == 2
+
+    # Test exception if structure not made of Blocks
+    with pytest.raises(TypeError):
+        _parse_structure([exp_factors], exp_factors)
+
+    # Test exception on missing factor level
+    fac_missing = {'fac1': ['a', 'b', 'c']}
+    tst_missing = [
+        Block(fac_missing, trials=10),
+        Block(exp_factors, trials=10)
+    ]
+    with pytest.raises(RuntimeError):
+        _parse_structure(tst_missing, exp_factors)
+
+    # Test exception on extra factor level
+    fac_extra = exp_factors.copy()
+    fac_extra['fac3'] = [200, 800]
+    tst_extra = [
+        Block(exp_factors, trials=10),
+        Block(fac_extra, trials=10)
+    ]
+    with pytest.raises(RuntimeError):
+        _parse_structure(tst_extra, exp_factors)
+
+    # Test exception when factors given but exp_factors is empty
+    with pytest.raises(RuntimeError):
+        _parse_structure(tst, {})